backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
google-cloud-aiplatform
google-cloud-bigquery
python-dotenv
pandas
